requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-pinecone>=0.0.1
python-dotenv>=1.0.0
pinecone-client>=3.0.0
psycopg2-binary>=2.9.9
pypdf>=3.17.0